Add base page for localization
