+++ tests/puppeteer/pages/BO/international/languages/index.js
@@ -1,16 +1,12 @@
 require('module-alias/register');
-const BOBasePage = require('@pages/BO/BObasePage');
+const LocalizationBasePage = require('@pages/BO/international/localization/localizationBasePage');
 
-module.exports = class Languages extends BOBasePage {
+module.exports = class Languages extends LocalizationBasePage {
   constructor(page) {
     super(page);
 
     this.pageTitle = 'Languages •';
 
-    // Header Selectors
-    this.localizationNavItemLink = '#subtab-AdminLocalization';
-    this.geolocationNavItemLink = '#subtab-AdminGeolocation';
-    this.currenciesNavItemLink = '#subtab-AdminCurrencies';
     // Selectors grid panel
     this.gridPanel = '#language_grid_panel';
     this.gridTable = '#language_grid_table';
@@ -24,31 +20,6 @@ module.exports = class Languages extends BOBasePage {
     this.tableRow = `${this.tableBody} tr:nth-child(%ROW)`;
     this.tableEmptyRow = `${this.tableBody} tr.empty_row`;
     this.tableColumn = `${this.tableRow} td.column-%COLUMN`;
-  }
-
-  /* Header methods */
-  /**
-   * Go to currencies page
-   * @return {Promise<void>}
-   */
-  async goToSubTabCurrencies() {
-    await this.clickAndWaitForNavigation(this.currenciesNavItemLink);
-  }
-
-  /**
-   * Go to localization page
-   * @return {Promise<void>}
-   */
-  async goToSubTabLocalization() {
-    await this.clickAndWaitForNavigation(this.localizationNavItemLink);
-  }
-
-  /**
-   * Go to geolocation page
-   * @return {Promise<void>}
-   */
-  async goToSubTabGeolocation() {
-    await this.clickAndWaitForNavigation(this.geolocationNavItemLink);
   }
 
   /* Reset methods */
