+++ tests/puppeteer/pages/BO/international/localization/index.js
@@ -1,38 +1,10 @@
 require('module-alias/register');
-const BOBasePage = require('@pages/BO/BObasePage');
+const LocalizationBasePage = require('@pages/BO/international/localization/localizationBasePage');
 
-module.exports = class Localization extends BOBasePage {
+module.exports = class Localization extends LocalizationBasePage {
   constructor(page) {
     super(page);
 
     this.pageTitle = 'Localization • ';
-    this.languagesNavItemLink = '#subtab-AdminLanguages';
-    this.currenciesNavItemLink = '#subtab-AdminCurrencies';
-    this.geolocationNavItemLink = '#subtab-AdminGeolocation';
-  }
-
-  /* Header Methods */
-  /**
-   * Go to languages page
-   * @return {Promise<void>}
-   */
-  async goToSubTabLanguages() {
-    await this.clickAndWaitForNavigation(this.languagesNavItemLink);
-  }
-
-  /**
-   * Go to currencies page
-   * @return {Promise<void>}
-   */
-  async goToSubTabCurrencies() {
-    await this.clickAndWaitForNavigation(this.currenciesNavItemLink);
-  }
-
-  /**
-   * Go to geolocation page
-   * @return {Promise<void>}
-   */
-  async goToSubTabGeolocation() {
-    await this.clickAndWaitForNavigation(this.geolocationNavItemLink);
   }
 };
